chore: Rename store to stores
