+++ src/app/login/page.tsx
@@ -2,10 +2,12 @@
 
 import LoginForm from '@components/login/LoginForm';
 
-export default function Login() {
+const LoginPage = () => {
   return (
     <main className="min-h-screen flex items-center justify-center bg-white dark:bg-gray-900 transition-colors">
       <LoginForm />
     </main>
   );
-}
+};
+
+export default LoginPage;
